+++ src/views/v3/Bridge/Routes/details/index.ts
@@ -1,0 +1,4 @@
+import MaxSlippage from './MaxSlippage';
+import MinOutput from './MinOutput';
+import RouteDetails from './RouteDetails';
+export default { MaxSlippage, MinOutput, RouteDetails };
